requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.24.0
pydantic==2.10.4
snowflake-connector-python==3.13.1
pandas==2.2.3
python-dotenv==1.0.0
scikit-learn==1.6.0
numpy==1.26.4
joblib~=1.5.1